requirements.txt
Flask
azure-storage==0.36.0
boto3==1.4.7